pyproject.toml
[tool.poetry]
name = "met-impression"
version = "0.1.0"
description = ""
authors = ["Chris Barry <2490974+ChrisW-B@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "met_impression"}]

[tool.poetry.scripts]
start = "met_impression.__main__:main"

[tool.poetry.dependencies]
python = "^3.7"
requests = "^2.28.1"
inky = {extras = ["rpi", "example-depends"], version = "^1.4.0"}
hitherdither = {git = "https://www.github.com/hbldh/hitherdither"}
Pillow = "^9.3.0"
spotipy = "^2.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
